Code Cleanup

+++ src/main/java/com/hoi4utils/clausewitz/map/seed/SeedProbabilityMap_GPU_3.java
@@ -102,7 +102,7 @@ public class SeedProbabilityMap_GPU_3 extends AbstractMapGeneration {
         double[] distanceModifiers = new double[r + 1]; // 0 to r
         distanceModifiers[0] = 0.0;
         for (int i = 1; i <= r; i++) {
-            distanceModifiers[i] = (double) (i - 0.2) / r; // Normalize the value to be between 0 and 1
+            distanceModifiers[i] = (i - 0.2) / r; // Normalize the value to be between 0 and 1
         }
 
         final int[] _data = new int[size];
@@ -258,7 +258,7 @@ public class SeedProbabilityMap_GPU_3 extends AbstractMapGeneration {
         double[] distanceModifiers = new double[r + 1]; // 0 to r
         distanceModifiers[0] = 0.0;
         for (int i = 1; i <= r; i++) {
-            distanceModifiers[i] = (double) (i - 0.2) / r; // Normalize the value to be between 0 and 1
+            distanceModifiers[i] = (i - 0.2) / r; // Normalize the value to be between 0 and 1
         }
 
         // int rSquared = r * r;
